Bump version to 0.3.0 and assign new UUID
This package used be called `DopplerDriftSearchPipelineOutputs`,
but now it it is called `DopplerDriftSerachTools`.  To allow the new
name to be registered, a new UUID is needed.  Assigning a new UUID
seemed like a big deal (though neither name has been registered in
`General`) so the version was also bumped to 0.3.0.